requirements.txt
torch
zetascale
swarms
einops 